Add processor for parser map generation
