+++ processor/src/main/java/com/workday/autoparse/xml/codegen/AutoparseProcessor.java
@@ -10,20 +10,12 @@ package com.workday.autoparse.xml.codegen;
 import com.workday.autoparse.xml.annotations.XmlAttribute;
 import com.workday.autoparse.xml.annotations.XmlChildElement;
 import com.workday.autoparse.xml.annotations.XmlElement;
-import com.workday.autoparse.xml.annotations.XmlParserPartition;
 import com.workday.autoparse.xml.annotations.XmlPostParse;
 import com.workday.autoparse.xml.annotations.XmlTextContent;
 import com.workday.autoparse.xml.annotations.XmlUnknownElement;
 import com.workday.autoparse.xml.utils.CollectionUtils;
-import com.workday.autoparse.xml.utils.StringUtils;
-import com.workday.meta.PackageTree;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
 import java.util.Set;
 
 import javax.annotation.processing.AbstractProcessor;
@@ -31,9 +23,7 @@ import javax.annotation.processing.RoundEnvironment;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.PackageElement;
 import javax.lang.model.element.TypeElement;
-import javax.lang.model.util.ElementFilter;
 import javax.tools.Diagnostic;
 
 /**
@@ -46,8 +36,6 @@ import javax.tools.Diagnostic;
  */
 public class AutoparseProcessor extends AbstractProcessor {
 
-    private Map<String, Collection<TypeElement>> parserMap = new HashMap<>();
-
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
         if (annotations == null || annotations.isEmpty()) {
@@ -58,7 +46,6 @@ public class AutoparseProcessor extends AbstractProcessor {
         Set<? extends Element> elements = roundEnv.getElementsAnnotatedWith(XmlElement.class);
         for (Element element : elements) {
             if (element.getKind() == ElementKind.CLASS) {
-                addClassToParseMap((TypeElement) element);
                 generateClassParser((TypeElement) element);
             }
         }
@@ -72,13 +59,6 @@ public class AutoparseProcessor extends AbstractProcessor {
             }
         }
 
-        // Generate ParserMaps
-        Set<PackageElement> partitionPackageElements = ElementFilter.packagesIn(
-                roundEnv.getElementsAnnotatedWith(XmlParserPartition.class));
-        PackageTree packageTree =
-                new PackageTree(processingEnv.getElementUtils(), partitionPackageElements);
-
-        generateParserMaps(packageTree);
         return true;
     }
 
@@ -89,31 +69,12 @@ public class AutoparseProcessor extends AbstractProcessor {
                                           XmlPostParse.class.getCanonicalName(),
                                           XmlChildElement.class.getCanonicalName(),
                                           XmlAttribute.class.getCanonicalName(),
-                                          XmlTextContent.class.getCanonicalName(),
-                                          XmlParserPartition.class.getCanonicalName());
+                                          XmlTextContent.class.getCanonicalName());
     }
 
     @Override
     public SourceVersion getSupportedSourceVersion() {
         return SourceVersion.latestSupported();
-    }
-
-    private void addClassToParseMap(TypeElement element) {
-        XmlElement annotation = element.getAnnotation(XmlElement.class);
-        for (String parseKey : annotation.value()) {
-            if (StringUtils.isNotEmpty(parseKey)) {
-                putInCollectionMap(parserMap, parseKey, element);
-            }
-        }
-    }
-
-    private <K, V> void putInCollectionMap(Map<K, Collection<V>> map, K key, V value) {
-        Collection<V> collection = map.get(key);
-        if (collection == null) {
-            collection = new ArrayList<>();
-            map.put(key, collection);
-        }
-        collection.add(value);
     }
 
     private void generateClassParser(TypeElement element) {
@@ -123,54 +84,6 @@ public class AutoparseProcessor extends AbstractProcessor {
             processingEnv.getMessager()
                          .printMessage(Diagnostic.Kind.ERROR, e.getMessage(), element);
         }
-
-    }
-
-    private void generateParserMaps(PackageTree packageTree) {
-        Map<PackageElement, Map<String, TypeElement>> mapsByPackage = splitParserMap(packageTree);
-        for (Map.Entry<PackageElement, Map<String, TypeElement>> entry : mapsByPackage.entrySet()) {
-            ParserMapGenerator generator =
-                    new ParserMapGenerator(processingEnv, entry.getKey(), entry.getValue());
-            try {
-                generator.generateParseMap();
-            } catch (IOException e) {
-                processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR, e.getMessage());
-            }
-        }
-    }
-
-    private Map<PackageElement, Map<String, TypeElement>> splitParserMap(PackageTree packageTree) {
-        Map<PackageElement, Map<String, TypeElement>> mapsByPackage = new HashMap<>();
-        for (Map.Entry<String, Collection<TypeElement>> entry : parserMap.entrySet()) {
-            String parseKey = entry.getKey();
-            for (TypeElement element : entry.getValue()) {
-                // A null matching package means that this element goes into the default partition,
-                // which is keyed by null here.
-                PackageElement matchingPackage = packageTree.getMatchingPackage(element);
-                Map<String, TypeElement> map = mapsByPackage.get(matchingPackage);
-                if (map == null) {
-                    map = new HashMap<>();
-                    mapsByPackage.put(matchingPackage, map);
-                }
-                TypeElement previousValue = map.put(parseKey, element);
-                if (previousValue != null) {
-                    String packageString = matchingPackage != null
-                                           ? String.format(Locale.US,
-                                                           "partition under package '%s'",
-                                                           matchingPackage.getQualifiedName())
-                                           : "the default partition";
-                    String errorMessage =
-                            String.format("%s and %s both tried to map to tag name \"%s\" in %s.",
-                                          element.getQualifiedName(),
-                                          previousValue.getQualifiedName(),
-                                          parseKey,
-                                          packageString);
-                    processingEnv.getMessager()
-                                 .printMessage(Diagnostic.Kind.ERROR, errorMessage, element);
-                }
-            }
-        }
-        return mapsByPackage;
     }
 
 }
