+++ processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,2 @@
 com.workday.autoparse.xml.codegen.AutoparseProcessor
+com.workday.autoparse.xml.codegen.XmlElementParserMapProcessor
